remove mvp uv drawing, separate uv rendering
