+++ Smash Forge/GUI/Menus/RenderSettings.cs
@@ -25,7 +25,6 @@ namespace Smash_Forge.GUI
 
             // Misc Settings
             renderCameraPathCB.Checked = Runtime.renderPath;
-            drawUvCB.Checked = Runtime.drawUv;
             textParamDir.Text = Runtime.paramDir;
             BackgroundGradient1.BackColor = Runtime.backgroundGradientTop;
             BackgroundGradient2.BackColor = Runtime.backgroundGradientBottom;
@@ -777,11 +776,6 @@ namespace Smash_Forge.GUI
                 Runtime.hitboxAnglesColor = Color.FromArgb(0xFF, colorDialog.Color);
                 pbHitboxAnglesColor.BackColor = Runtime.hitboxAnglesColor;
             }
-        }
-
-        private void drawUvCB_CheckedChanged(object sender, EventArgs e)
-        {
-            Runtime.drawUv = drawUvCB.Checked;
         }
 
         private void swagYCB_CheckedChanged(object sender, EventArgs e)
